rename import for getOrCreateDialogue
